+++ crates/ui/src/plot/shape/pie.rs
@@ -1,15 +1,15 @@
 // @reference: https://d3js.org/d3-shape/pie
 
-use std::f64::consts::TAU;
+use std::f32::consts::TAU;
 
 use super::arc::ArcData;
 
 #[allow(clippy::type_complexity)]
 pub struct Pie<T> {
-    value: Box<dyn Fn(&T) -> Option<f64>>,
-    start_angle: f64,
-    end_angle: f64,
-    pad_angle: f64,
+    value: Box<dyn Fn(&T) -> Option<f32>>,
+    start_angle: f32,
+    end_angle: f32,
+    pad_angle: f32,
 }
 
 impl<T> Default for Pie<T> {
@@ -31,26 +31,26 @@ impl<T> Pie<T> {
     /// Set the value of the Pie.
     pub fn value<F>(mut self, value: F) -> Self
     where
-        F: 'static + Fn(&T) -> Option<f64>,
+        F: 'static + Fn(&T) -> Option<f32>,
     {
         self.value = Box::new(value);
         self
     }
 
     /// Set the start angle of the Pie.
-    pub fn start_angle(mut self, start_angle: f64) -> Self {
+    pub fn start_angle(mut self, start_angle: f32) -> Self {
         self.start_angle = start_angle;
         self
     }
 
     /// Set the end angle of the Pie.
-    pub fn end_angle(mut self, end_angle: f64) -> Self {
+    pub fn end_angle(mut self, end_angle: f32) -> Self {
         self.end_angle = end_angle;
         self
     }
 
     /// Set the pad angle of the Pie.
-    pub fn pad_angle(mut self, pad_angle: f64) -> Self {
+    pub fn pad_angle(mut self, pad_angle: f32) -> Self {
         self.pad_angle = pad_angle;
         self
     }
